requirements.txt
django
dj-database-url
python-dotenv
whitenoise
django-crispy-forms
crispy-bootstrap5
django-taggit
openai
gunicorn
psycopg2-binary
markdown
Pillow
